Only add size of non whitelisted files and directories

+++ src/main/java/in/ashwanthkumar/gocd/actions/DeleteAction.java
@@ -16,7 +16,7 @@ import java.util.Set;
 public class DeleteAction implements Action {
     private static final Logger LOG = LoggerFactory.getLogger(DeleteAction.class);
 
-    Set<String> whiteList = new HashSet<>();
+    private Set<String> whiteList = new HashSet<>();
 
     public DeleteAction(Set<String> whiteList) {
         this.whiteList = whiteList;
@@ -29,24 +29,53 @@ public class DeleteAction implements Action {
     @Override
     public long invoke(File pipelineDir, String version, boolean dryRun) {
         File versionDir = new File(pipelineDir.getAbsolutePath() + "/" + version);
-        long size = FileUtils.sizeOfDirectory(versionDir);
+        DirectoryStats stats = getDirectoryStats(versionDir);
 
-        if (dryRun) {
-            LOG.info("[DRY RUN] Will remove " + versionDir.getAbsolutePath() + ", size = " + FileUtils.byteCountToDisplaySize(size));
-        } else {
-            LOG.info("Deleting " + versionDir.getAbsolutePath() + ", size = " + FileUtils.byteCountToDisplaySize(size));
-            try {
-                deleteDirectory(versionDir);
-            } catch (IOException e) {
-                LOG.error(e.getMessage(), e);
-                throw new RuntimeException(e);
+        if (stats.fileCount > 0) {
+            String path = versionDir.getAbsolutePath();
+            String displaySize = FileUtils.byteCountToDisplaySize(stats.size);
+
+            if (dryRun) {
+                LOG.info("[DRY RUN] Will remove " + path + ", size = " + displaySize);
+            }
+            else {
+                LOG.info("Deleting " + path + ", size = " + displaySize);
+
+                try {
+                    deleteDirectory(versionDir);
+                } catch (IOException e) {
+                    LOG.error(e.getMessage(), e);
+                    throw new RuntimeException(e);
+                }
             }
         }
 
-        return size;
+        return stats.size;
     }
 
-    void deleteDirectory(File path) throws IOException {
+    private DirectoryStats getDirectoryStats(File directory) {
+        DirectoryStats stats = new DirectoryStats();
+
+        File[] files = directory.listFiles();
+
+        if(files != null) {
+            for (File file : files) {
+                try {
+                    if (!FileUtils.isSymlink(file) && isNotWhiteListed(file)) {
+                        if (file.isDirectory())
+                            stats.Add(getDirectoryStats(file));
+                        else
+                            stats.Add(file);
+                    }
+                }
+                catch (IOException ignored) {}
+            }
+        }
+
+        return stats;
+    }
+
+    private void deleteDirectory(File path) throws IOException {
         if (isNotWhiteListed(path)) {
             File[] files = path.listFiles();
             if (files == null) throw new IOException("Couldn't list files inside " + path.getAbsolutePath());
@@ -69,4 +98,18 @@ public class DeleteAction implements Action {
         return !this.whiteList.contains(path.getName());
     }
 
+    private class DirectoryStats {
+        private long size = 0;
+        private long fileCount = 0;
+
+        void Add(DirectoryStats directoryStats) {
+            size += directoryStats.size;
+            fileCount += directoryStats.fileCount;
+        }
+
+        void Add(File file) {
+            size += file.length();
+            fileCount += 1;
+        }
+    }
 }
